bot/requirements.txt
python-telegram-bot[job-queue]~=22.0
pydantic-settings~=2.0
pyyaml~=6.0
